sdk: Add missing parameter in code generator
Some calls to the __extendsSubCollection method are missing on of
the mandatory parameters. This code might have never been executed
before, but it executes when trying to regenerate from the latest
version of the API schema.

Also changed the default connection parameters (host, port, user
and password) to match defaults used by the engine, as this
simplifies the use of the code generator.

Also fixed a minor typo.

Change-Id: I03e7c8b05d260d2f16239c71dd21daa4cd6cb495
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>