sdk: Remove trailing slashes from URL prefix
Currently we get the URL prefix of the RESTAPI from the URL provided by
the user, without modification. If this prefix contains trailing slashes we
end up building complete URLs that contain two slashes. For example, if
the user writes this code:

  api = ovirtsdk.api.API(url='https://whatever/api/', ...)
  api.clusters.list()

We send the following request to the server:

  GET /api//clusters HTTP/1.1

The server doesn't accept these two slashes and generates an error
response.

To fix this issue this patch removes trailing slashes from the URL
prefix.

Change-Id: I7a75aeccd7ba42fa3d49c4cebb9fc0ab3eff5cb9
Bug-Url: https://bugzilla.redhat.com/1063746
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>