sdk: Check SSL server name
Currently we don't check that the host name provided in the URL matches
the host name contained in the server certificate. This is a common
feature of most SSL clients, but it isn't well supported by the SSL
implementation in Python 2.6. To improve security this patch explicitly
checks the host name given in the URL against the subject common name
attribute and the subject alternative names extension. This check will
be enabled by default and disabled when using "insecure=True" in the
constructor of the entry point object.

Change-Id: I6a08179288fd564fbb79f5731c3e32251e560d81
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>