sdk: No new lines in basic auth token
Currently we generate the value of the basic authentication header using
the "encodestring" method of the "base64" module, but the output of this
method may contain many new lines, depending on the length of the user
name and password. When more than one newline is included we only delete
the last, and as a result we generate a header with a new line, which is
interpreted by the web server as two headers, thus breaking the
authentication mechanism. This patch changes the SDK so that all the new
lines are removed.

Change-Id: I02bb66b5bed9c152fcba95ab577373b66f19be05
Bug-Url: https://bugzilla.redhat.com/1085837
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>