codegen: Sort subcollection fields
The code generator doesn't sort the fields containing references to
subcollections. This means that the generated code may be different
depending on the default order of dictionaries, or the order of
collections in the XML schema and RSDL files. To avoid potential
uncessary changes this patch changes the code generator so that the
fields are generated in alphabetical order.

Change-Id: Ib4d9af8282d7e7fc0b9fb06e85c1dad1b62355ff
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit e76524ba08ca5c860773829bdbb0266c6ec23cdc)