codegen: Allow = in replaced function parameters
The code generator replaces some of the functions in the params.py file
generated by generateDS.py. To locate these functions it uses regular
expressions, but it doesn't take into account that parameters can
contain the equals sign. In version 2.9a of generateDS.py the generated
functions didn't have parameters with default values, so the equals sign
wasn't present:

  def parseString(inString)

In version 2.12a of generateDS.py the generated functions have
parameters with default values:

  def parseString(inString, silence=False)

This patch changes the code generator so that it will locate the
functions correctly in both cases, with or without the equals sign.

Change-Id: I7eccef883def440e04e47a7e04c476a3d3348a91
Related: https://bugzilla.redhat.com/1091688
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>