codegen: Check version of generateDS.py
Currently we don't check the version of the generateDS.py, as a
consecuence the generated code can be different depending on the machine
where the code generator is executed. This patch adds a check of the
version of generateDS.py, so if isn't the correct one the generation
will fail.

Change-Id: Iddd3cb4a814db111dc125fbfdf6481d276354517
Related: https://bugzilla.redhat.com/1091688
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>