codegen: Support plural exceptions
Currently we have a set of entity names whose corresponding collection
name isn't formed by just adding an "s" to the end: the name of the
entity and the collection is the same. But we will soon introduce some
entities where the name of the collection can't be derived in this
simple way, in particular we will introduce the SchedulingPolicy entity
and the SchedulingPolicies collection. To support that, this patch
changes the list of collections to a map, so that the name of the entity
and the name of the collection can be different.

Change-Id: I97d45d533e12723426f98dad2874de3945fd483b
Related: https://bugzilla.redhat.com/1114655
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>