codegen: Ignore errors when cleaning old connection
The constructor of the API object tries to clean connections left around
by the user, calling their "disconnect" method. If this fails the
creation of the new API object fails as well. To avoid that this patch
changes the SDK so that any error during that cleaup is ignored.

Change-Id: Id2a5c690aa9336bc0b441ca57100b9d6cd09e335
Bug-Url: https://bugzilla.redhat.com/1116009
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 3b6ed1ed646eb57eadacc59e0e66cfa56d4547b2)