sdk: Calculate singulars correctly
There are a few places in the code generator where the singular form of
words is calculated directly removing the last character, assuming
it will be an "s". This isn't correct, as there are exceptions. The
correct way is to call the StringUtils.toSingular() method.

A previous patch fixed that in the code generator. This patch updates
the generated code. That removes an incorrectly generated
"SchedulingPolicie" class.

Change-Id: I1c1a8665d9d49ac496ab13cea9ce08a45d906184
Related: https://bugzilla.redhat.com/1114655
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>