sdk: Fix file names in MANIFEST.in
A previous commit changed the format of the documentation files from
plain text to markdown, and added the .md suffix to the file names, but
the MANIFEST.in file wasn't according correctly. As a result the new
documentation files aren't included in the source release. This patch
fixes the MANIFEST.in file.

Change-Id: I8ecf51a2f49b6b7fed53528f1ef1030c349adad9
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>