codegen: Apply filter by name/alias after get
Currently, for searchable collections, the "get" method is implemented
using the "search" query parameter, and returning to the caller all the
results returned by the server. But when this is combined with the
"Filter: true" header the server ignores the "search" query parameter,
and instead it returns all the objects that the user has permissions
for. This means that in some situations the user may be requesting an
object and getting back a different one. To avoid this issue this patch
adds an additional filtering step, so the collection returned by the
server is filtered, and only the objects matching the name given by the
user are taken into consideration.

Change-Id: Ia762356ee4ad12839e089de2cd60f832bd212af5
Related-To: https://bugzilla.redhat.com/1155678
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 6f076f7fb76373c5fc64cf90687877f83318925b)