sdk: Apply filter by name/alias after get
Currently, for searchable collections, the "get" method is implemented
using the "search" query parameter, and returning to the caller all the
results returned by the server. But when this is combined with the
"Filter: true" header the server ignores the "search" query parameter,
and instead it returns all the objects that the user has permissions
for. This means that in some situations the user may be requesting an
object and getting back a different one. To avoid this issue this patch
adds an additional filtering step, so the collection returned by the
server is filtered, and only the objects matching the name given by the
user are taken into consideration.

Change-Id: I9baf549c68a93f329b76d41a2ce52b7a7e7b6c79
Bug-Url: https://bugzilla.redhat.com/1155678
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>