codegen: Extract metadata from Maven artifacts
Currently we require a live engine to generate the metadata. This
increases the time required to re-generate the SDK, specially when there
are more than one version that needs to be re-generated. This patch
changes the code generator so that it extracts the metadata from the
engine Maven artifacts, thus no live engine is required.

Change-Id: I059364a47d3194074ad5dcd373acf7bd3f3ce0e7
Bug-Url: https://bugzilla.redhat.com/1069204
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>