sdk: Manually fix scheduling policy naming issue
The code generator assumes that the name of the URL segment for
scheduling policies is "schedulingpolicy", but in fact it is
"schedulingpolicie". This is a bug of the server side. This patch fixes
it manually to reduce the diffs with the code generated with the next
version of the code generator.

Change-Id: I8e044c1831191811af13fd49641bc5b828222d88
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>