sdk: Manually fix type of disk snapshots
Currently the RSDL generator of the engine doesn't report correctly the
type of disk snapshots collection, it reports "BaseResource" instead of
"DiskSnapshot". The current version of the generator doesn't have a
problem with this, because it doesn't use this information (it
calculates it from the collection type, using plural to singular rules).
But the new version of the generator will stick to the reported
information, and will generate wrong code. To avoid this issue this
patch fixes the type manually. The issue will be fixed in the engine by
the following change:

  http://gerrit.ovirt.org/37755

Change-Id: I1e2d046c9f10f5bd7576671ec29b31e1a2799ce5
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>