sdk: Manually fix brick parameter name
This name of the parameter used to add a Gluster brick is currently
"glusterbricks", and that is wrong, it should be "brick". This happens
because the current version of the code generator calculates types using
plural to singular (and the other way around) rules, instead of getting
the exact information from the XML schema. This issue will be properly
fixed in the next version of the generator. This patch fixes it manually
to reduce the differences on the generated code.

Change-Id: I090119f6d3972970ecf81756139f57a21ebd4769
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>