sdk: Manually fix VM NUMA node broker class
This class extends the "NumaNode" parameter class, and that is wrong, it
should extend "VirtualNumaNode". This happens because the current
version of the code generator calculates types using plural to singular
(and the other way around) rules, instead of getting the exact
information from the XML schema. This issue will be properly fixed in
the next version of the generator. This patch fixes it manually to
reduce the differences on the generated code.

Change-Id: I35f5c82a7b3d592f9570f500b8684eeda2494203
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>