sdk: Manually fix host storage broker class
This class extends the "HostStorage" parameter class, and that is wrong,
it should extend "Storage". This happens because the current version of
the code generator calculates types using plural to singular (and the
other way around) rules, instead of getting the exact information from
the XML schema. This issue will be properly fixed in the next version of
the generator. This patch fixes it manually to reduce the differences on
the generated code.

Change-Id: I8cc05bb01ce2267fb4ee1bd5ee77ace1e2f4d6f5
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>