codegen: Avoid calculation of singulars and plurals
Currently the code generator assumes that there is a semantic
relationship between the names of XML schema complex type that represent
collections and entities: the name of the entity is the plural of the
name of the collection. To handle this it uses rules to calculate the
plurals from singulars, and the other way around. But for some
collections and entities simple calculation of plural and singular
(adding or removing a trailing "s") doesn't work. For example:

  SchedulingPolicies -> SchedulingPolicy

Another example is the following, which has been introduced in the
server side:

  KatelloErrata -> KatelloErratum

The code generator also assumes that the names of the XML schema complex
types match the URL segments. This doesn't always work. For example,
with the recent addition of external providers to the server side there
are URL segments like this one:

  /externalhostproviders/{externalhostprovider:id}/hosts

The XML complex type corresponding to this "hosts" collection is
"ExternalHost", and not "Host" as the existing rules expect.

We have been handling these situations adding exceptions in the code of
the generator, and implementing more robust rules for calculations of
plurals and singulars, but this doesn't scale well, and it is prone to
unexcpected failures. To avoid these issues in a more systematic way
this patch modifies the generator so that it will extract the required
information from the XML schema and the RSDL instead of making
calculations.

The core of this change is the introduction of a tree data structure
that stores the links extracted from the RSDL, so that it is easy to
find from one node its children and parents. This data structure is used
to decide if a location corresponds to a collection, entity or action,
and combined with querying the XML schema to extract the type names.

Change-Id: I7b27061d436e524683dbed81c5521eed51e8c2f0
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>