codegen, sdk: Fix name of "parentclass" parameter
A previous change in the code generator induced a change in the names of
some of the the "parentclass" parameters of entity brokers. For example,
the constructor of the "ClusterAffinityGroupVM" broker class changed
from this:

  def __init__(self, clusteraffinitygroup, vm, context):

To this:

  def __init__(self, affinitygroup, vm, context):

This change shouldn't affect any client of the SDK, because this is an
implementation detail of the broker classes, and clients shouldn't rely
on this, but the CLI does rely, as it uses the parameter name to figure
out the type name of the parent broker:

  clusteraffinitygroup -> ClusterAffinityGroup

The result is that the CLI doesn't find the correct parent class, and
that means that autocompletion doesn't work correctly.

This patch reverts to the previous behaviour so that the CLI will work
as expected.

Change-Id: I36f6033eed5fb0a83964743e61e4072c21f07dee
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>