sdk: Always generate the tarball when building RPMs
Currently when building a RPM the tarball is generated using the "git
archive" command and the HEAD tree. This means that changes that haven't
been locally commited aren't included in the result. In addition the
tarball is only generated if there were changes in the Makefile or the
.spec template, which has the same effect. To avoid these two issues
this patch changes the command used to build the tarball and also
changes the Makefile so that the tarball is always regenerated.

Change-Id: Ib16253e18f39ff4bafef1fcdebe21e8f456b4b23
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>