sdk: Always set the request body
The cURL library remembers the body of the request, and this means that
after sending a POST with a body the next GET will have the same body.
This isn't harmful, as servers will typically ignore bodies of GET
requests, but it isn't ideal. To solve this issue this patch changes the
SDK so that it will explicitly clear the request body.

Change-Id: I028f2672fa7f817b21c12f8805d5bf09ae2bb61e
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>