sdk: Decode headers using ASCII
Currently the HTTP response headers are used without decoding them.
This is fine in Python 2 because strins are actually array of bytes, but
in Python 3 strings aren't array of bytes, so before using the headers
then need to be decoded using the right encoding. In this case the right
encoding is ASCII, as it is what is recommended by the HTTP protocol
specification.

Change-Id: Iee51e67addd7ba7a53c726c5ab8d2270cd348392
Related: https://bugzilla.redhat.com/1096137
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>