sdk: Avoid altering dictionary while removing elements
Currently the "clear" method of the "Cache" class iterates the keys of
the dictionary and removes them, but this causes errors in Python 3.
This patch changes that methods so that it makes a copy of the list of
keys and iterates it intead of the original one.

Change-Id: Iff458d9e432c47e4c5d9ef7ec3e1178d718be497
Related: https://bugzilla.redhat.com/1096137
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>