sdk: Don't add headers without value
Currently the SDK adds to the request all the headers passed from the
generated broker classes, including those headers with no value. The
result is headers that have the string "None" as the value, for example:

  Content-type: None
  Correlation-Id: None

These headers, in particular the "Content-type: None" header are then
rejected by the application server.

This behaviour is a regression introduced by the change tht replaced the
Pyhton HTTP module with PycURL, and only affects the 3.6 branck.

To avoid this issue this patch changes the SDK so that it will not add
the headers that don't have a value.

Change-Id: I1ebbfc2197bbfdaf3765460b3ff68765065b7696
Bug-Url: https://bugzilla.redhat.com/1227655
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>