sdk: Read complete HTTP reason code
Currently the SDK doesn't read the complete HTTP reason code sent by the
server, only the first word. For example, if the server sends the
following response line:

  HTTP/1.1 503 Service Unavailable

The SDK will only get the first word and then it will generate an
incomplete error report:

  status: 503
  reason: Service

This patch fixes that problem, so that the error report will contain the
complete HTTP reason code:

  status: 503
  reason: Service Unavailable

This is a regression that was introduced as part of the changes to use
PycURL instead of the builtin Python HTTP module.

Change-Id: Ic7fff3d5b5c81aa949985a085afaf67c437365c9
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>