packaging: Distribution specific RPM spec files
Currently we use a single .spec file template for all the supported
distributions. To support all the distributions it should contain many
"%if" statements that change behaviour, for example to build or not the
Python 3 support packages. This would make it convoluted and difficult
to maintain. To avoid this issue this patch introduces different spec
file templates, one for each supported distribution.

Change-Id: Id9d733bdfcafab4e07325695ff395a40a9b27bd2
Bug-Url: https://bugzilla.redhat.com/1260865
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>