sdk: Support URL parameters in actions
Currently the SDK doesn't support the URL parameters that are available
in some actions. For example, the action to deactivate a host supports
the "async" parameter, but it isn't possible to use in the SDK because
the signature of the "deactivate" method doesn't contain it:

  def deactivate(
    self,
    action=params.Action(),
    correlation_id=None
  )

It should be like this instead:

  def deactivate(
    self,
    action=params.Action(),
    correlation_id=None,
    async=None
  )

In order to support that this patch modifies the code generator so that
these URL parameters are added to the signature, and sent to the server.

Change-Id: I7cad178d3b7d809d5bb81efcc128a671e97bdea3
Bug-Url: https://bugzilla.redhat.com/1311495
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>