Add support for debug log file
Currently when the debug mode is enabled the SDK writes the debug output
to stdout. This isn't convenient for applications that need to write
other things to stdout. This patch adds a new "log" parameter to the
constructor of the connection. This parameter can contain a file name or
an IO object. If it contains a file name the file will be created if
needed, and the debug output will be added to the end of the file.  If it
contains an IO object then the debug output will be written directly to
that IO object.

Change-Id: I0abd38f103290b23df193052a6fab42843cc2928
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>