Use only query parameters
Currently the SDK has to send parameters as both query and matrix
parameters, as the server some times requires query and some times
matrix. But the server has recently been fixed to support both, for all
paramers, so the SDK no longer needs to send both. This patch changes
the SDK so that it only uses query parameters.

Change-Id: I41f0aa2a0bb4b8b20e71a64769a352ea508c7583
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>