Move connection class to "ovirtsdk4" module
Currently the "Connection" and "ConnectionBuilder" classes are part of
the "ovirtsdk4.http" module, but then the classes are imported from the
"ovitsdk4" module so that the user can just "import ovirtsdk4". But this
has the downside that tools that automatically generate documentation
will generate it under "ovirtsdk4.http", which can be confusing for
users. To avoid this potential source of confusion this patch moves the
"Connection" and "ConnnectionBuilder" classes to the "__init__.py" file
of the "ovirtsdk4" module.

Change-Id: Ia79b93b2e510a39af1592e0886ee958f5e90f5a7
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>