sdk: Use __module__ to check module membership
The SDK needs to check if the parameters given to constructors of
classes in the "params" module are brokers, as they have a special
treatment. In order to do that it checks if they belong to the brokers
module, using the "ReflectionHelper.isModuleMember" function. This
function iterates the members of the module in order to do the check. A
side effect of that is that the module needs to be loaded, and an
exception is generated if it isn't. To avoid that this patch changes the
implementation of the "isModuleMember" function so that it uses a
simpler check, using the "__module__" builtin to obtain and compare the
module name.

Change-Id: I1f10343659b1fd440e521b9981f1b359aa4fe050
Bug-Url: https://bugzilla.redhat.com/1333889
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>