Add support for following links to lists
Currently the "follow_link" method of the "Conection" class only
supports links to single objects, but not links to list of objects. This
patch changes the SDK so that it will also support this kind of links.
For example, when a VM is retrieved it will include links to its NICs:

<vm id="123" href="/vms/123">
  <link rel="nics" href="/vms/123/nics"/>
  ...
</vm>

The SDK will process this "link" element and store it in the "vm.nics"
property of the created object. Later, the user will be able to use the
"follow_link" method to retrieve that list:

vm = ...
nics = connection.follow_link(vm.nics)
for nic in nics:
    print("nic: %s" % nic.name)

Change-Id: I1a928cb2031318b5e8a1bc74ac2f23fe268cecbe
Signed-off-by: Ondra Machacek <omachace@redhat.com>
(cherry picked from commit a5f6e0713a29abe3ad9868a2aa1f83f07e6eadfa)