Use type name as the tag for method parameters
Currently the name of the XML tag used to represent the input parameters
of methods like "add" is the name of the parameter as described in the
model. For example, for the operation to add a disk attachment the name
of the parameter is "attachment", so the XML document generated and sent
by the SDK is something like this:

  <attachment>
    ...
  </attachment>

But the server doesn't expect this, it expects the name the XML tag
corresponding to the type of the parameter, in this case this case
"DiskAttachment", so the server expects the following:

  <disk_attachment>
    ...
  </disk_attachment>

It would be more correct to use the name of the parameter, as described
in the model, but it isn't factible to introduce such change in the
server side at this point, so we need to change the SDK to generate what
the server expects. That is what this patch does.

In the future, we will probably change the server so that it ignores the
root tag of the documents, then this patch can be reverted.

Change-Id: I18284685f5eefc981fd6769ac7e1ca998d586a81
Signed-off-by: Ondra Machacek <omachace@redhat.com>