Name plular element tag same as the member
Previously when we send action and contruct XML with
list of the entity we named tags by the entity. This
patch changes the behavior, so we name it by the member
which is written.

For example if we want to create bonds, we will send following xml:

  <action>
    <modified_bonds>
      <host_nic>
        <bonding>
          <options>...</options>
          <slaves>
            <host_nic>
              <name>eth0</name>
            </host_nic>
            <host_nic>
              <name>eth1</name>
            </host_nic>
          </slaves>
        </bonding>
        <name>bond0</name>
      </host_nic>
    </modified_bonds>
  </action>

Previously we sent <host_nics> instead of <modified_bonds>.

Change-Id: Id1405d540f95abc8dc9dcfa0ac3b2a477b4cba06
Signed-off-by: Ondra Machacek <omachace@redhat.com>