Add generic reader
Currently the SDK doesn't provide any mechanism to read an XML object
without explicitly specifying its type. For example, to read a virtual
machine the only method available is "VmReader.read_one". But there are
situations where the SDK can't know in advance what type of object is
going to be received from the server, for example "action" or "fault".
To support that, this patch adds a new "Reader.read" method that decides
the type of object based on the name of the first tag, and invokes the
corresponding specific reader to read it.

Change-Id: Ibe55914453ea6c6c7e46406bdcbad7acaa813721
Signed-off-by: Ondra Machacek <omachace@redhat.com>
(cherry picked from commit 4f8e4c5ba72b7b0777b372cc12d8cb1e4691a52f)