Make sure that query parameters are always sorted
Currently when we format query parameters we depend on the internal sorting of
the Python 'dict' type. This isn't convenient, specially for tets, as it isn't
possible to predict what will be the query string generated. To avoid that this
patch changes the SDK so that it sorts the query parameters by name before
creating the query string.

Change-Id: I19028dad41872633a82d8cfece9625e782b9d222
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>