Catch Unicode encode and decode exceptions
When an incorrect ASCII string is passed to the 'encode' method it generates an
'UnicodeDecodeError' exception and we catch it. But in Python 3 it generates
'UnicodeEncodeError', and we don't catch it. As a result we don't handle this
situation correctly in Python 3. This patch changes the SDK so that it catches
both exceptions, so that it will work with Python 2 and 3.

Change-Id: I9d5da5fcd9a6e8de51acd1aaa00a63797700617f
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>