Encode/decode test data using UTF-8
Currently the data sent and received by the tests isn't converted from strings
to arrays of bytes. As a result the tests don't work with Python 3 because the
HTTP server used by the tests requires bytes, not strings. This patch changes
the tests so that they explicitly encode/decode the data.

Change-Id: If7aad9db77ad12ebcd1d995060f4fddf820bcdfb
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>