Fix reading of lists of elements
Currently the method that read lists of elements doesn't handle
correctly empty elements without close tags. For example, when reading
the following cluster:

  <cluster>
    ...
    <required_rng_sources/>
    <switch_type>legacy</switch_type>
  </cluster>

The 'legacy' value of the 'switch_type' element is interpreted as one of
the values of the previous 'required_rng_sources' list.

This patch fixes the method and introduces two tests to verify that it
works correctly.

Change-Id: I76c3625719c5e2ef327fcee7cec974d740d5f655
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>