Empty list is not written to XML
When the user wants to send an empty list to the server, it is not
currently possible. For example when the user wants to create non VM
network he has to send an empty usages element. But if he sets the
"Network.usages" attribute to empty list, the empty usages element isn't
sent.

This patch fixes that.

Change-Id: I66bc71cf5c4757fff57b3143218488315dc218b8
Signed-off-by: Ondra Machacek <omachace@redhat.com>
(cherry picked from commit 19a123cbf8d3daa70f99432f9c469eae90a6beb2)