Accept 200 code for 'Add' methods
Currently the SDK doesn't accept the HTTP 200 code as a valid result of
'Add' methods, but some methods in the server return this code, for
example the method that adds a cluster quota limit. This patch changes
the SDK so that accepts it.

Change-Id: I59e2a5748bd4fc914a33d0431eacca6cdae543b9
Signed-off-by: Ondra Machacek <omachace@redhat.com>
(cherry picked from commit 33ec13f1bb1124674b4007f91b0c57cc0b97bb8f)