Avoid duplicate symbols in Mac OS X
The C compiler used in Mac OS X, clang, is more strict than the one used in
Linux, gcc. It detects that some symbols are duplicated, and the build fails.
This patch fixes the error that causes that failure, declaring the symbols as
'extern' in the .h file, and re-declaring them in the corresponding .c file.

Change-Id: Ice0331395266b05f963c334e14d7c9448301655b
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 1363be1e5db37ef7523088815a6fe0cb3e935570)