Clarify the units of `provisioned_size`
This patch changes the example to add a VM disk so that it is clearer
that size of a disk has to be specified in bytes. It also changes the
example to use a more realistic size, 10 GiB, as it currently the size
is 1 MiB, which isn't realistic.

Change-Id: I6a80abef018a3d58dcccfb4f99e4d18ba7d7c6ef
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
Signed-off-by: Ondra Machacek <omachace@redhat.com>
(cherry picked from commit 196961295b3f4b549067b1c2cab659da594edd69)