Fix writing 'id' attribute for enum values
Currently the SDK doesn't generate correctly the XML for attributes
named 'id' that contain enum values. For example, when writing the list
of SSO methods of a virtual machine it generates the following:

  <sso>
    <methods>
      <method>
        <id>guest_agent</id>
      </method>
    </methods>
  </sso>

But it should write the following instead, as the 'id' model attribute
should always be represented as an XML attribute:

  <sso>
    <methods>
      <method id="guest_agent"/>
    </methods>
  </sso>

This patch addresses that issue, which was mostly a typo in the
generator of the XML writing code.

Change-Id: I52065c9aeba4628155cbdc638bf0334834f4a15d
Bug-Url: https://bugzilla.redhat.com/1408839
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>