Automatically generate 'version.py'
Currently the 'version.py' file is updated manually with each release of
the project, and it needs to be in sync with the version specified in
the POM files. To avoid that this patch changes the SDK so that the
'version.py' file is atumatically generated, taking the version number
from the root POM file.

Change-Id: Ib9d5dce6b48c2e6c1a3dcc43fa17204c88096632
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>