Merge link with element if it has same name
This patch merges the link and element if it has same name.

For example when API returns following XML:

 <root>
  <link href='/link' rel'attrs'/>
  <attrs>
    <attr id='123'/>
  </attrs>
 </root>

We must ensure that both the link and element are read and merged into
single attribute of Python SDK, so user can later access the 'href' of
the link as well as content of the 'attr' element.

Change-Id: Ied8cdd24632083967cbc5c4d5528c8c1cf02ab63
Signed-off-by: Ondra Machacek <omachace@redhat.com>
(cherry picked from commit eb50d0cb724a14e43981846ad10c37bc6325d50a)