Don't generate custom tarball
Currently the SDK generates a tarball using the 'tar' command
explicitly. This tarball is different to the one generated by the
distutils 'sdist' command (which is also executed as part of the build).
To avoid differences between the files published in the oVirt web site
and the files published in PyPI, this patch changes the automation
scripts so that this custom tarball is replaced by the tarball generated
by the 'sdist' command. In addition, and in order to achieve that, the
patch also changes the way that the version numbers of the SDK are
generated, so that they follow the advices of PEP440. Note that these
changes in version numbers only apply to non release builds, release
builds already used PEP440 version numbers.

Change-Id: I8febc38acb3980ea78229ba384f1140273866a86
Bug-Url: https://bugzilla.redhat.com/1422979
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>